+++ apple-codesign/CHANGELOG.rst
@@ -5,8 +5,8 @@
 0.10.0
 ======
 
-* Support for stapling XAR archives (``.pkg`` files). (There is still no support for
-  signing XAR archives.)
+* Support for notarizing and stapling flat packages (``.pkg`` installers). (There
+  is still no support for signing flat packages.)
 
 0.9.0
 =====